requirements.txt
# Core dependencies
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
pyyaml>=6.0

# libclang for C++ parsing
libclang>=18.1.1

# Azure OpenAI
openai>=1.42.0
pydantic>=2.8.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Development
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
mypy>=1.0.0
